gfx: rgxkm: Fix screen tearing regression
Revert "rgx: fix tearing because of incorrect FlushDC" and apply IMG's
fix.

This patch is provided by IMG via file pp92376.diff from
IMG ticket 92376 - "Patch fix screen tearing regression"

The root cause of the screen tearing regression is believed to be that
a lock-up is incorrectly detected due to a race condition between the
FW starting to process a Client CCB after APM, and the WDT checking
the RGX Health Status.  This occurs in the case where a Client CCB was
previously empty; the host then adds a command to the CCB, powers on,
and kicks RGX. There is a small window between the FW updating its
write and read pointers.

This patch avoids the race condition by adding a condition to the
Client CCB lock-up detection such that it will not trigger, if on the
previous check that Client CCB was empty.  This targets the specific
race condition, allowing all the other Health Status checks to run
as before.

Bug: 19111950
BZ: 231394
Change-Id: I5a394ed3d86b69a17c2b7ab3588986db692045c5
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>