gpu: Move the intel_media drivers under staging.
Disable the EXTERNAL_INTEL_MEDIA subsystem. Replace it with a cleaner
implementation under the staging directory. Due to kbuild limitations
some files had to be renamed.

This change is substantial, but does not modify any C/H files at all,
so there should be no change in behavior.

This change is mostly to aid bisection of DDK issues after the upgrade
to DDK 1.6. These changes will be required for that DDK version.

Because no code changes are being made here, all of the Intel DDK
customizations remain intact.

Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>
Bug: 26592113
Change-Id: I08ce8bb1f34863c24e5ca6567cd6cf13d192360a