net: wireless: bcmdhd: Fix up the BRCM wifi DHD code
Security Vulnerability fix for memory overflow wifi driver
function wl_cfgvendor_rtt_set_config. In the current fix added
check to validate if the target_cnt is valid or not if it is not valid
then parse error. Since target_cnt can be controlled by user netlink
input which needs to validated at the DHD level.

Bug: 32219255

Change-Id: I7ee4992712d94ed832d65cd878cc56be9181cfee
Signed-off-by: Ecco Park <eccopark@google.com>